ecommerce_rest/requirements.txt
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-simple-history==3.0.0
djangorestframework==3.13.1
drf-spectacular==0.21.1
drf-spectacular-sidecar==2021.12.13
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.3.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
requests==2.27.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.7.0
